Moved countries & commodities endpoints under public API namespace
