+++ app/controllers/api/public/countries_controller.rb
@@ -1,20 +1,20 @@
 module Api
-  module V3
+  module Public
     class CountriesController < ApiController
       include Api::V3::ParamHelpers
 
       skip_before_action :load_context
-      before_action :set_collection, only: %i[index]
+      before_action :set_collection, only: [:index]
 
       def index
         render json: @collection,
-               each_serializer: Api::V3::Dashboards::CountrySerializer
+               each_serializer: Api::Public::CountrySerializer
       end
 
       private
 
       def set_collection
-        @collection = Api::V3::FilterCountries.new(filter_params).call
+        @collection = Api::Public::FilterCountries.new(filter_params).call
       end
 
       def filter_params
