+++ app/serializers/api/public/country_serializer.rb
@@ -1,0 +1,7 @@
+module Api
+  module Public
+    class CountrySerializer < ActiveModel::Serializer
+      attributes :id, :name
+    end
+  end
+end
