--- app/serializers/api/v3/dashboards/commodity_serializer.rb
@@ -1,14 +1,0 @@
-module Api
-  module V3
-    module Dashboards
-      class CommoditySerializer < ActiveModel::Serializer
-        attributes :id, :name
-        attribute :country_ids, if: :include_countries?
-
-        def include_countries?
-          @instance_options[:include_countries]
-        end
-      end
-    end
-  end
-end
