--- app/serializers/api/v3/dashboards/country_serializer.rb
@@ -1,9 +1,0 @@
-module Api
-  module V3
-    module Dashboards
-      class CountrySerializer < ActiveModel::Serializer
-        attributes :id, :name
-      end
-    end
-  end
-end
